% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Simplified simulator runner}
\usage{
run_simulation(simulator, epochs, calculate.pcf = FALSE, pcf_grid)
}
\arguments{
\item{simulator}{1-species simulator object (1d, 2d or 3d)}

\item{epochs}{amount of population-dependent time units to run}

\item{calculate.pcf}{if TRUE, adds pcf estimate to output}

\item{pcf_grid}{grid for pcf calculation}
}
\value{
list with results
}
\description{
Simplified simulator runner
}
