% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_run_simulations.R
\name{batch_run_simulations}
\alias{batch_run_simulations}
\title{Runs multiple simulations}
\usage{
batch_run_simulations(
  area_length_x,
  b = 1,
  d = 0,
  dd,
  sw,
  sm = 1,
  work_dir = file.path(getwd(), Sys.time()),
  calculate.pcf = FALSE,
  pcf.r = 5 * pmax(sm, sw),
  pcf.nodes = 101,
  ndim = 1,
  periodic = TRUE,
  seed = 12345,
  initial_population = 1000,
  realtime_limit = 3600,
  epochs = 1000,
  birth_tail_cutoff = 1e-06,
  death_radius_cutoff = 5 * sw,
  kernel_nodes = 101,
  cell_count_x = 100,
  n_cores = availableCores()
)
}
\arguments{
\item{area_length_x}{}

\item{b}{}

\item{d}{}

\item{dd}{}

\item{sw}{}

\item{sm}{}

\item{work_dir}{}

\item{calculate.pcf}{}

\item{pcf.r}{}

\item{pcf.nodes}{}

\item{ndim}{}

\item{periodic}{}

\item{seed}{}

\item{initial_population}{}

\item{realtime_limit}{}

\item{epochs}{}

\item{birth_tail_cutoff}{}

\item{death_radius_cutoff}{}

\item{kernel_nodes}{}

\item{cell_count_x}{}

\item{n_cores}{}
}
\value{

}
\description{
Experimental
}
